Refactor of TC_13_51_01
